requirements.txt
telethon>=1.32
pika>=1.3
psycopg2-binary>=2.9
redis>=5.0
pyyaml>=6.0
python-dotenv>=1.0
pytest>=7.4
